While I waited for the msg.match bug to be fixed, I added a bunch a stuff :)
