+++ .jscsrc
@@ -1,0 +1,4 @@
+{
+    "preset": "google",
+    "validateIndentation": 4
+}
