+++ .jshintrc
@@ -1,0 +1,29 @@
+{
+  "asi": false,
+  "expr": true,
+  "loopfunc": true,
+  "strict": false,
+  "curly": true,
+  "evil": true,
+  "curly": true,
+  "white": true,
+  "trailing": true,
+  "eqeqeq": true,
+  "immed": true,
+  "indent": 4,
+  "debug": true,
+  "newcap": true,
+  "undef": true,
+  "unused": false,
+
+  "quotmark": "single",
+  "maxdepth": 3,
+  "maxlen": 80,
+
+  "node": true,
+  "browser": true,
+
+  "globals": {
+    "window": true
+  }
+}
